requirements.txt
anyio==4.9.0
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.4.26
charset-normalizer==3.4.1
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dulwich==0.22.8
factory_boy==3.3.3
Faker==37.1.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
more-itertools==10.7.0
msgpack==1.1.0
packaging==25.0
pbs-installer==2025.4.9
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
PyJWT==2.9.0
pyproject_hooks==1.2.0
pytest==8.3.5
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.4.28.22
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
zstandard==0.23.0
